Refactor profile.py into multiple files
Improve maintainability by splitting the profile logic up in separate
files based on the profile version thus creating smaller files. Also
move them to a separate folder.
