+++ app/models/profiles/exceptions.py
@@ -1,0 +1,14 @@
+class XMLNotValidError(Exception):
+    pass
+
+
+class GraphParseError(Exception):
+    pass
+
+
+class GraphNotConformError(Exception):
+    pass
+
+
+class ProfileVersionRetiredError(Exception):
+    pass
