+++ app/models/profiles/profile.py
@@ -1,0 +1,320 @@
+from abc import ABC, abstractmethod
+from pathlib import Path
+
+from jinja2 import Environment, FileSystemLoader, select_autoescape
+from lxml import etree
+from rdflib import Graph, URIRef
+from pyshacl import validate as shacl_validate
+from pysparql_anything import SparqlAnything
+
+from app.models.profiles.exceptions import (
+    GraphNotConformError,
+    GraphParseError,
+    XMLNotValidError,
+)
+
+sa = SparqlAnything()
+
+
+class Profile(ABC):
+    def __init__(self, bag_path: Path):
+        self.bag_path = bag_path
+        self.jinja_env = Environment(
+            loader=FileSystemLoader(self.query_sip().parent),
+            autoescape=select_autoescape(),
+        )
+
+    @staticmethod
+    @abstractmethod
+    def query_sip() -> Path:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def query_descriptive_ie() -> Path:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def query_premis_ie() -> Path:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def query_premis_representation() -> Path:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def shacl_sip() -> Path:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def shacl_profile() -> Path:
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def profile_name() -> str:
+        """The name of the profile, as defined in the SIP spec."""
+        pass
+
+    def _validate_premis(self) -> list[XMLNotValidError]:
+        """Validate the PREMIS files.
+
+        There are multiple premis files, one on the package level and
+        one for each representation.
+
+        Returns:
+            A list with errors detailing the parse/validation errors.
+        """
+        errors = []
+
+        premis_xsd = etree.XMLSchema(etree.parse(self.premis_xsd()))
+
+        premis_package_path: Path = self.bag_path.joinpath(
+            "data", "metadata", "preservation", "premis.xml"
+        )
+
+        # PREMIS on package level
+        try:
+            premis_package = etree.parse(premis_package_path)
+            premis_xsd.assertValid(premis_package)
+        except (etree.DocumentInvalid, etree.ParseError) as e:
+            errors.append(XMLNotValidError(str(e.error_log)))
+
+        # PREMIS on representation level
+        for premis_representation_path in self.bag_path.glob(
+            str(
+                Path(
+                    "data",
+                    "representations",
+                    "representation_*",
+                    "metadata",
+                    "preservation",
+                    "premis.xml",
+                )
+            )
+        ):
+            try:
+                premis_representation = etree.parse(premis_representation_path)
+                premis_xsd.assertValid(premis_representation)
+            except (etree.DocumentInvalid, etree.ParseError) as e:
+                errors.append(XMLNotValidError(str(e.error_log)))
+
+        return errors
+
+    def _validate_mets(self) -> list[XMLNotValidError]:
+        """Validate the METS files.
+
+        There are multiple METS files, one on the package level and one for
+        every representation.
+
+        Returns:
+            A list with errors detailing the parse/validation errors.
+        """
+        mets_xsd = etree.XMLSchema(etree.parse(self.mets_xsd()))
+
+        mets_package_path: Path = self.bag_path.joinpath("data", "mets.xml")
+
+        mets_representation_path: Path = self.bag_path.joinpath(
+            "data", "representations", "representation_1", "mets.xml"
+        )
+
+        errors = []
+        # METS on package level
+        try:
+            mets_package = etree.parse(mets_package_path)
+            mets_xsd.assertValid(mets_package)
+        except (etree.DocumentInvalid, etree.ParseError) as e:
+            errors.append(XMLNotValidError(str(e.error_log)))
+
+        # METS on representation level
+        for mets_representation_path in self.bag_path.glob(
+            str(Path("data", "representations", "representation_*", "mets.xml"))
+        ):
+            try:
+                mets_representation = etree.parse(mets_representation_path)
+                mets_xsd.assertValid(mets_representation)
+            except (etree.DocumentInvalid, etree.ParseError) as e:
+                errors.append(XMLNotValidError(str(e.error_log)))
+
+        return errors
+
+    def validate_metadata(self) -> list[XMLNotValidError]:
+        """Validate the metadata files.
+
+        A profile has:
+            - Multiple METS files (package and one per representation level).
+            - Multiple PREMIS files (package one per and representation level).
+            - One or more descriptive metadata file (one on package and possible
+              one per representation level).
+
+        Returns:
+            A list with errors detailing the parse/validation errors.
+        """
+        errors: list = self._validate_premis()
+        errors.extend(self._validate_mets())
+        errors.extend(self._validate_descriptive())
+
+        return errors
+
+    @abstractmethod
+    def _validate_descriptive(self) -> list[XMLNotValidError]:
+        """Validate the descriptive metadata.
+
+        Returns:
+            A list with errors detailing the parse/validation errors.
+        """
+        pass
+
+    @abstractmethod
+    def _construct_shacl_graph(self) -> Graph:
+        """Construct a graph containing the SHACLs.
+
+        This is used for validating the data graph.
+
+        Returns:
+            A SHACL graph.
+        """
+
+        pass
+
+    def validate_graph(self, data_graph: Graph) -> bool:
+        """Validate if the graph is conform.
+
+        Returns: True if the graph was conform.
+
+        Raises:
+            GraphNotConformError: If not conform, containing the reason why.
+        """
+        # An empty graph conforms with the SHACL. Check if the graph has at least
+        # a premis:intellecutalEntity node.
+        if (
+            None,
+            URIRef("http://www.w3.org/1999/02/22-rdf-syntax-ns#type"),
+            URIRef("http://www.loc.gov/premis/rdf/v3/IntellectualEntity"),
+        ) not in data_graph:
+            raise GraphNotConformError(
+                "Graph is perceived as empty as it does not contain an intellectual entity."
+            )
+        shacl_graph = self._construct_shacl_graph()
+        conforms, results_graph, results_text = shacl_validate(
+            data_graph=data_graph,
+            shacl_graph=shacl_graph,
+            meta_shacl=True,
+            allow_warnings=True,
+        )
+
+        if not conforms:
+            raise GraphNotConformError(results_text)
+
+        return True
+
+
+class BasicProfile(Profile):
+
+    def _validate_descriptive(self) -> list[XMLNotValidError]:
+        """Validate the dcterms file.
+
+        Basic profile has one file with the descriptive metadata.
+
+        Returns:
+            A parse/validation error in a list if applicable.
+        """
+        dcterms_xsd = etree.XMLSchema(etree.parse(self.dcterms_xsd()))
+
+        dcterms_package_path: Path = self.bag_path.joinpath(
+            "data",
+            "metadata",
+            "descriptive",
+            "dc_1.xml",
+        )
+
+        # DCTERMS on package level
+        try:
+            dcterms_package = etree.parse(dcterms_package_path)
+            dcterms_xsd.assertValid(dcterms_package)
+        except (etree.DocumentInvalid, etree.ParseError) as e:
+            return [XMLNotValidError(str(e.error_log))]
+
+        return []
+
+    def parse_graph(self) -> Graph:
+        """Parse the metadata as a graph.
+
+        As the SPARQL service paths are dynamic, we need to instantiate
+        the SPARQL file and write it in the bag folder.
+
+        Transform the metadata to rdf and then load it into a graph.
+
+        Returns:
+            The graph.
+        Raises:
+            GraphParseError: If parsing failed.
+        """
+        # write SPARQL-anything query for the SIP.
+        query_sip_destination = self.bag_path.joinpath(self.query_sip().name)
+        sip_template = self.jinja_env.get_template(str(self.query_sip().name))
+        with open(str(query_sip_destination), "w") as f:
+            f.write(sip_template.render(bag_path=self.bag_path))
+
+        # write SPARQL-anything query for the descriptive metadata on IE level.
+        query_descriptive_ie_destination = self.bag_path.joinpath(
+            self.query_descriptive_ie().name
+        )
+        profile_template = self.jinja_env.get_template(
+            str(self.query_descriptive_ie().name)
+        )
+        with open(str(query_descriptive_ie_destination), "w") as f:
+            f.write(profile_template.render(bag_path=self.bag_path))
+
+        # write SPARQL-anything query for the premis metadata on the IE level.
+        query_premis_ie_destination = self.bag_path.joinpath(
+            self.query_premis_ie().name
+        )
+        profile_template = self.jinja_env.get_template(str(self.query_premis_ie().name))
+        with open(str(query_premis_ie_destination), "w") as f:
+            f.write(profile_template.render(bag_path=self.bag_path))
+
+        # write SPARQL-anything query for the premis metadata on the representation level.
+        query_premis_rep_destination = self.bag_path.joinpath(
+            self.query_premis_representation().name
+        )
+        profile_template = self.jinja_env.get_template(
+            str(self.query_premis_representation().name)
+        )
+        with open(str(query_premis_rep_destination), "w") as f:
+            f.write(profile_template.render(bag_path=self.bag_path))
+
+        # Run SPARQL-anything transformation.
+        try:
+            sip_graph = sa.construct(q=str(query_sip_destination), f="TTL")
+            profile_descriptive_ie = sa.construct(
+                q=str(query_descriptive_ie_destination), f="TTL"
+            )
+            profile_premis_ie = sa.construct(
+                q=str(query_premis_ie_destination), f="TTL"
+            )
+            profile_premis_rep = sa.construct(
+                q=str(query_premis_rep_destination), f="TTL"
+            )
+
+            data_graph = (
+                sip_graph
+                + profile_descriptive_ie
+                + profile_premis_ie
+                + profile_premis_rep
+            )
+
+        except Exception as e:
+            raise GraphParseError(f"Error when parsing graph: {str(e)}")
+
+        return data_graph
+
+    def _construct_shacl_graph(self) -> Graph:
+        shacl_graph = Graph()
+        shacl_graph.parse(str(self.shacl_sip()), format="turtle")
+        shacl_graph.parse(str(self.shacl_profile()), format="turtle")
+        return shacl_graph
